% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createSTMGrid.R
\name{createSTMGrid}
\alias{createSTMGrid}
\title{Generate the grid of chip centers}
\usage{
createSTMGrid(waferRay, dieWidth, dieHeigth, originalDieCenterX,
  originalDieCenterY)
}
\arguments{
\item{dieWidth:}{the width of a chip}

\item{dieHeight:}{the height of a chip}

\item{waferRay:}{the ray of the wafer}

\item{originalDieCenterX:}{the x coordinate of the (0,0) chip}

\item{originalDieCentery:}{the y coordinate of the (0,0) chip}
}
\value{
a list made of the x axes and the y axes.
}
\description{
This function creates the grid of all the chip centers in a wafer from
basic informations about the wafer
}

