% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCoordsFromLogical.R
\name{getCoordsFromLogical}
\alias{getCoordsFromLogical}
\title{Turn locical to absolute coordinates}
\usage{
getCoordsFromLogical(logicalCoords, dieWidth, dieHeigth, originalDieCenterX,
  originalDieCenterY)
}
\arguments{
\item{dieWidth:}{the width of a chip}

\item{dieHeight:}{the height of a chip}

\item{logicalCoords:}{a matrix containing a couple of coordinates in each line}

\item{originalDieCenterX:}{the x coordinate of the (0,0) chip}

\item{originalDieCentery:}{the y coordinate of the (0,0) chip}
}
\value{
a matrix whose rows are (x,y) absolute coordinats. The origin is the center of the wafer.
}
\description{
This function turns logical coordinates in the absolute position
of the center of a chip.
}

