% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getLogicalCoordsFromData.R
\name{getLogicalCoordsFromData}
\alias{getLogicalCoordsFromData}
\title{Extract defect coordiantes}
\usage{
getLogicalCoordsFromData(waferData)
}
\arguments{
\item{waferData:}{the data.frame containing all the defect data}
}
\value{
a two-columns matrix where each row represent a pair of (x,y) coordinates.
}
\description{
This function extracts the logical defect coordiantes form a properly formatted data.frame
}

