% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexFromCoords.R
\name{indexFromCoord}
\alias{indexFromCoord}
\title{Turn fault coordinates in indexes.}
\usage{
indexFromCoord(coords, features, grid)
}
\arguments{
\item{coords}{a matrix containing a couple of coordinates in each line}

\item{features}{a list containing the dimensional features of the wafer. Required fields are:
dieWidth: the width of a chip
dieHeight: the height of a chip
waferRay: the ray of the wafer}

\item{grid}{a list of two vectors:
x: x coordinates of the centers of the chips on the wafer.
y: y coordinates of the centers of the chips on the wafer.}
}
\value{
a matrix whose rows are (i,j) index of faulty chips.
}
\description{
This function turns fault coordinates in indexes to fill a fault map.
}

