% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readExcel.R
\name{readExcel}
\alias{readExcel}
\title{Read an Excel file}
\usage{
readExcel(xlsToRead, saveOnRDS = TRUE, rdsToWrite = "xlsData.rds", ...)
}
\arguments{
\item{xlsToRead:}{the path of the Excel file to be read.}

\item{saveOnRDS:}{a logical flag. If set to TRUE, the returned data frame
is saved to a file, too. Default is FALSE.}

\item{rdsToWrite:}{if "saveOnRDS" is TRUE, the path of the rds file where the data
should be saved.
If nothing is specified, a defaut file "xlsData.rds" is used.}

\item{...:}{argument to be passed to lower functions.}
}
\value{
a data.frame containing the data in the Excel file.
}
\description{
This function loads an Excel file which is assumed to contain a data frame
and, optionally, saves it on a file.
}

