% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiftReference.R
\name{shiftReference}
\alias{shiftReference}
\title{Shift the orign of the reference to the left lower corner}
\usage{
shiftReference(logicalCoords, originalDieCenterX, originalDieCenterY, grid)
}
\arguments{
\item{logicalCoords:}{a matrix containing a couple of logical coordinates in each line -
ie: the fist chip is 0, the following 1, the previous -1, ...}

\item{grid:}{the positions of all the centers of the chips in the wafer}

\item{originalDieCenterX:}{the x coordinate of the (0,0) chip}

\item{originalDieCentery:}{the y coordinate of the (0,0) chip}
}
\value{
a matrix whose rows are logical coordinates. The origin is the left lower corner of the wafer.
}
\description{
This function turns logical coordinates in the absolute position
of the center of a chip.
}

