% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createSTMGrid.R
\name{createSTMGrid}
\alias{createSTMGrid}
\title{Generate the grid of chip centers}
\usage{
createSTMGrid(waferRay, dieWidth, dieHeight, originalDieCenterX = 0,
  originalDieCenterY = 0)
}
\arguments{
\item{waferRay}{the ray of the wafer}

\item{dieWidth}{the width of a chip}

\item{dieHeight}{the height of a chip}

\item{originalDieCenterX}{the x coordinate of the (0,0) chip. Default is 0.}

\item{originalDieCenterY}{the y coordinate of the (0,0) chip  Default is 0.}
}
\value{
a list made of the x axis and the y axis.
}
\description{
This function creates the grid of all the chip centers in a wafer from
basic informations about the wafer.
}

