% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fillMatrixFromIndex.R
\name{fillMatrixFromIndex}
\alias{fillMatrixFromIndex}
\title{Fill a matrix according to parameters}
\usage{
fillMatrixFromIndex(index, matchingValue, nonMatchingValue, nrow, ncol)
}
\arguments{
\item{index}{a two-column matrix containing a couple of index in each line -
these are assumed to be index of a matrix, so they have to be positive integers.}

\item{matchingValue}{the value to put in any element of the matrix matching one of
the coordinates in index.}

\item{nonMatchingValue}{the value to put in any element of the matrix not matching one of
the coordinates in index.}

\item{nrow}{number of rows of the returned matrix.}

\item{ncol}{number of columns of the returned matrix.}
}
\value{
a matrix filled as requested.
}
\description{
This function fills a matrix of simbolic values according to the index passed
as a parameter.
}

